Updated package name to npcvillagers (from .example)
