--- src/main/java/com/example/npcvillage/configs/WebSecurityConfig.java
@@ -1,4 +1,0 @@
-package com.example.npcvillage.configs;
-
-public class WebSecurityConfig {
-}
