--- src/main/java/com/example/npcvillage/controllers/NpcController.java
@@ -1,4 +1,0 @@
-package com.example.npcvillage.controllers;
-
-public class NpcController {
-}
