--- src/main/java/com/example/npcvillage/models/AppUser.java
@@ -1,5 +1,0 @@
-package com.example.npcvillage.models;
-
-public class AppUser {
-
-}
