--- src/main/java/com/example/npcvillage/models/Npc.java
@@ -1,4 +1,0 @@
-package com.example.npcvillage.models;
-
-public class Npc {
-}
