--- src/main/java/com/example/npcvillage/models/enums/Species.java
@@ -1,4 +1,0 @@
-package com.example.npcvillage.models.enums;
-
-public enum Species {
-}
