--- src/main/java/com/example/npcvillage/repos/AppUserRepository.java
@@ -1,4 +1,0 @@
-package com.example.npcvillage.repos;
-
-public interface AppUserRepository {
-}
