--- src/main/java/com/example/npcvillage/repos/NpcRepository.java
@@ -1,4 +1,0 @@
-package com.example.npcvillage.repos;
-
-public interface NpcRepository {
-}
