--- src/main/java/com/example/npcvillage/services/OpenAiApiHandler.java
@@ -1,4 +1,0 @@
-package com.example.npcvillage.services;
-
-public class OpenAiApiHandler {
-}
