--- src/main/java/com/example/npcvillage/services/UserDetailsServiceImpl.java
@@ -1,4 +1,0 @@
-package com.example.npcvillage.services;
-
-public class UserDetailsServiceImpl {
-}
