+++ src/main/java/com/npcvillagers/npcvillage/configs/WebSecurityConfig.java
@@ -1,0 +1,4 @@
+package com.npcvillagers.npcvillage.configs;
+
+public class WebSecurityConfig {
+}
