+++ src/main/java/com/npcvillagers/npcvillage/controllers/AppUserController.java
@@ -1,0 +1,4 @@
+package com.npcvillagers.npcvillage.controllers;
+
+public class AppUserController {
+}
