+++ src/main/java/com/npcvillagers/npcvillage/controllers/NpcController.java
@@ -1,0 +1,4 @@
+package com.npcvillagers.npcvillage.controllers;
+
+public class NpcController {
+}
