+++ src/main/java/com/npcvillagers/npcvillage/models/AppUser.java
@@ -1,0 +1,5 @@
+package com.npcvillagers.npcvillage.models;
+
+public class AppUser {
+
+}
