+++ src/main/java/com/npcvillagers/npcvillage/models/Npc.java
@@ -1,0 +1,4 @@
+package com.npcvillagers.npcvillage.models;
+
+public class Npc {
+}
