+++ src/main/java/com/npcvillagers/npcvillage/models/enums/Alignment.java
@@ -1,0 +1,4 @@
+package com.npcvillagers.npcvillage.models.enums;
+
+public enum Alignment {
+}
