+++ src/main/java/com/npcvillagers/npcvillage/repos/AppUserRepository.java
@@ -1,0 +1,4 @@
+package com.npcvillagers.npcvillage.repos;
+
+public interface AppUserRepository {
+}
