+++ src/main/java/com/npcvillagers/npcvillage/services/OpenAiApiHandler.java
@@ -1,0 +1,4 @@
+package com.npcvillagers.npcvillage.services;
+
+public class OpenAiApiHandler {
+}
