+++ src/main/java/com/npcvillagers/npcvillage/services/UserDetailsServiceImpl.java
@@ -1,0 +1,4 @@
+package com.npcvillagers.npcvillage.services;
+
+public class UserDetailsServiceImpl {
+}
